requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
google-generativeai>=0.5.3
Werkzeug==2.3.7
pymongo>=4.0.0
requests>=2.31.0
anthropic>=0.7.0
chromadb>=0.4.0
